Code refactor - assignment and task
