+++ Requests.http
@@ -16,30 +16,34 @@ Accept: application/json
 }
 
 ### assign a task to a user
-POST http://localhost:8080/tasks/assign
+POST http://localhost:8080/assignments/
 Content-Type: application/json
 Accept: application/json
 
-{"taskId": 1, "userId":666, "deadline": "2022-06-19"}
+{"taskId": 3, "userId":666, "deadline": "2022-06-19"}
 
 ### change deadline in an assignment
-PUT http://localhost:8080/tasks/assign
+PUT http://localhost:8080/assignments
 Content-Type: application/json
 Accept: application/json
 
-{"taskId": 1, "userId":666, "deadline": "2025-12-31"}
+{
+  "taskId": 3,
+  "userId":666,
+  "deadline": "2025-12-31"
+}
 
 ### delete assignment
-DELETE http://localhost:8080/tasks/assign/666/1
+DELETE http://localhost:8080/assignments/666/3
 Content-Type: application/json
 Accept: application/json
 
 ### delete task and all associated assignments
-DELETE http://localhost:8080/tasks/1
+DELETE http://localhost:8080/tasks/2
 Content-Type: application/json
 Accept: application/json
 
 ### get assignment info
-GET http://localhost:8080/tasks/assignments/666/1
+GET http://localhost:8080/assignments/666/3
 Content-Type: application/json
 Accept: application/json
