+++ src/main/java/com/lft/taskservice/tasks/adapters/persistance/JpaAssignmentRepositoryAdapter.java
@@ -1,0 +1,56 @@
+package com.lft.taskservice.tasks.adapters.persistance;
+
+import com.lft.taskservice.tasks.adapters.logging.TaskLogging;
+import com.lft.taskservice.tasks.domain.Assignment;
+import com.lft.taskservice.tasks.domain.Task;
+import com.lft.taskservice.tasks.ports.AssignmentRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
+import java.time.LocalDate;
+import java.util.List;
+
+@Transactional
+@Repository
+@RequiredArgsConstructor
+public class JpaAssignmentRepositoryAdapter implements AssignmentRepository {
+
+    private final JpaAssignmentRepository assignmentRepository;
+    private final JpaAssignmentMapper assignmentMapper;
+    private final JpaTaskMapper taskMapper;
+
+    @Override
+    public void deleteAllAssignmentsToGivenTask(Long taskId) {
+        assignmentRepository.deleteAllAssignmentsAssociatedWithTaskId(taskId);
+    }
+
+    @TaskLogging
+    @Override
+    public Assignment assignTaskToUser(Assignment assignment) {
+        var assignmentEntity = assignmentMapper.toEntity(assignment);
+        var savedAssignmentEntity = assignmentRepository.save(assignmentEntity);
+        return assignmentMapper.toDomain(savedAssignmentEntity);
+    }
+
+    @TaskLogging
+    @Override
+    public List<Task> getAllTasksAssignedToUser(Long userId) {
+        return taskMapper.mapList(assignmentRepository.findAllTasksAssignedToUser(userId), Task.class);
+    }
+
+    @Override
+    public void changeDeadline(Long userId, Long taskId, LocalDate deadline) {
+        assignmentRepository.updateDeadline(deadline, taskId, userId);
+    }
+
+    @Override
+    public void deleteAssignment(Long userId, Long taskId) {
+        assignmentRepository.deleteAssignment(userId, taskId);
+    }
+
+    @Override
+    public Assignment getInfoOnAssignment(Long userId, Long taskId) {
+        return assignmentMapper.toDomain(assignmentRepository.findByUserIdAndTaskId(userId, taskId));
+    }
+}
