+++ src/main/java/com/lft/taskservice/tasks/adapters/persistance/JpaTaskRepositoryAdapter.java
@@ -1,15 +1,12 @@
 package com.lft.taskservice.tasks.adapters.persistance;
 
 import com.lft.taskservice.tasks.adapters.logging.TaskLogging;
-import com.lft.taskservice.tasks.domain.Assignment;
 import com.lft.taskservice.tasks.domain.Task;
 import com.lft.taskservice.tasks.ports.TaskRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 
 import javax.transaction.Transactional;
-import java.time.LocalDate;
-import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
@@ -20,10 +17,9 @@ public class JpaTaskRepositoryAdapter implements TaskRepository {
 
     private final JpaTaskRepository taskRepository;
     private final JpaTaskMapper taskMapper;
-    private final JpaAssignmentRepository assignmentRepository;
-    private final JpaAssignmentMapper assignmentMapper;
 
     @TaskLogging
+    @Override
     public Task save(Task task) {
         var taskEntity = taskMapper.toEntity(task);
         var savedTaskEntity = taskRepository.save(taskEntity);
@@ -31,6 +27,7 @@ public class JpaTaskRepositoryAdapter implements TaskRepository {
     }
 
     @TaskLogging
+    @Override
     public Task findById(Long id) {
       var taskEntity = Optional.of(taskRepository.findById(id).get()).orElseThrow(NoSuchElementException::new);
       return taskMapper.toDomain(taskEntity);
@@ -41,39 +38,4 @@ public class JpaTaskRepositoryAdapter implements TaskRepository {
     public void deleteTaskById(Long taskId) {
         taskRepository.deleteById(taskId);
     }
-
-    @Override
-    public void deleteAllAssignmentsToGivenTask(Long taskId) {
-        assignmentRepository.deleteAllAssignmentsAssociatedWithTaskId(taskId);
-    }
-
-    @TaskLogging
-    public Assignment assignTaskToUser(Assignment assignment) {
-        var assignmentEntity = assignmentMapper.toEntity(assignment);
-        var savedAssignmentEntity = assignmentRepository.save(assignmentEntity);
-        return assignmentMapper.toDomain(savedAssignmentEntity);
-    }
-
-    @TaskLogging
-    @Override
-    public List<Task> getAllTasksAssignedToUser(Long userId) {
-        return taskMapper.mapList(assignmentRepository.findAllTasksAssignedToUser(userId), Task.class);
-    }
-
-    @Override
-    public void changeDeadline(Long userId, Long taskId, LocalDate deadline) {
-        assignmentRepository.updateDeadline(deadline, taskId, userId);
-    }
-
-    @Override
-    public void deleteAssignment(Long userId, Long taskId) {
-        assignmentRepository.deleteAssignment(userId, taskId);
-    }
-
-    @Override
-    public Assignment getInfoOnAssignment(Long userId, Long taskId) {
-        return assignmentMapper.toDomain(assignmentRepository.findByUserIdAndTaskId(userId, taskId));
-    }
-
-
 }
