+++ src/main/java/com/lft/taskservice/tasks/ports/AssignmentService.java
@@ -1,0 +1,20 @@
+package com.lft.taskservice.tasks.ports;
+
+import com.lft.taskservice.tasks.domain.Assignment;
+import com.lft.taskservice.tasks.domain.Task;
+
+import java.util.List;
+
+public interface AssignmentService {
+
+    Assignment assignTaskToUser(Assignment assignment);
+
+    List<Task> getAllTasksAssignedToUser(Long userId);
+
+    void changeDeadline(Assignment assignment);
+
+    void deleteAssignment(Long taskId, Long userId);
+
+    Assignment getInfoOnAssignment(Long userId, Long taskId);
+
+}
