+++ src/main/java/com/lft/taskservice/tasks/ports/TaskService.java
@@ -1,25 +1,12 @@
 package com.lft.taskservice.tasks.ports;
 
-import com.lft.taskservice.tasks.domain.Assignment;
 import com.lft.taskservice.tasks.domain.Task;
-
-import java.util.List;
 
 public interface TaskService {
 
     Task save(Task task);
 
-    Assignment assignTaskToUser(Assignment assignment);
-
-    List<Task> getAllTasksAssignedToUser(Long userId);
-
-    void changeDeadline(Assignment assignment);
-
-    void deleteAssignment(Long taskId, Long userId);
-
     Task findTaskById(Long taskId);
-
-    Assignment getInfoOnAssignment(Long userId, Long taskId);
 
     void deleteTaskById(Long taskId);
